🧑‍💻 added Repository and refactored Code
