+++ articles/src/main/java/de/restockoffice/article/DataImport.java
@@ -1,8 +1,8 @@
-package de.restockoffice;
+package de.restockoffice.article;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
+import de.restockoffice.exceptions.DataImportException;
 import io.quarkus.runtime.StartupEvent;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.Observes;
@@ -20,22 +20,25 @@ public class DataImport {
     @Inject
     ObjectMapper mapper;
 
+    @Inject
+    ArticleRepository articleRepository;
+
     private static final Logger log = LoggerFactory.getLogger(DataImport.class);
 
     @Transactional
     public void loadInitialData(@Observes StartupEvent ev) {
-        if (PanacheEntityBase.count() > 0) return;
+        if (articleRepository.count() > 0) return;
 
         try (InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream("products.json")) {
             if (is == null) {
-                throw new RuntimeException("products.json nicht im Classpath gefunden!");
+                throw new DataImportException("Die Datei products.json wurde nicht im Classpath gefunden!");
             }
-            List<Article> articles = mapper.readValue(is, new TypeReference<List<Article>>() {});
-            PanacheEntityBase.persist(articles);
+            List<Article> articles = mapper.readValue(is, new TypeReference<List<Article>>() {
+            });
+            articleRepository.persist(articles);
             log.info("{} Articles importet successfully", articles.size());
         } catch (Exception e) {
             log.error("Exception in importing Articles: {}", e.getMessage());
-            e.printStackTrace();
         }
     }
 }
