+++ articles/src/main/java/de/restockoffice/exceptions/DataImportException.java
@@ -1,0 +1,11 @@
+package de.restockoffice.exceptions;
+
+public class DataImportException extends RuntimeException{
+    public DataImportException(String message) {
+        super(message);
+    }
+
+    public DataImportException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
